Data/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.9
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.1
scikit-surprise==1.1.0
scipy==1.4.1
seaborn==0.10.1
shell==1.0.1
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.17
sqlparse==0.3.0
surprise==0.1
threadpoolctl==2.1.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
